requirements.txt
fastapi
uvicorn
aiosmtplib
python-dotenv
python-multipart
email-validator
resend